move process check to its own file
